+++ R/dtq.R
@@ -13,14 +13,14 @@ NULL
 #' @details \emph{timestamp} represents the time of query (or chain) log after it is evaluated.
 #' @export
 dtl <- function(print = FALSE, chain = FALSE, purge = FALSE){
-  env <- dtq_id <- . <- query <- elapsed <- in_rows <- out_rows <- NULL
+  src <- env <- dtq_id <- . <- query <- elapsed <- in_rows <- out_rows <- NULL
   if(isTRUE(purge)) return(invisible(.DTQ$purge()))
   dt <- .DTQ$process()
   if(!isTRUE(chain)){
     if(isTRUE(print)) dt <- dt[, .SD, .SDcols=-c("dtq","dtq_call")]
   }
   else if(isTRUE(chain)){
-    dt <- dt[,.(dtq_depth=.N, query=paste(query,collapse=""), timestamp=timestamp[.N], elapsed=sum(elapsed), in_rows=in_rows[1L], out_rows=out_rows[.N]), .(dtq_id, env)]
+    dt <- dt[,.(dtq_depth=.N, query=paste(query, collapse=""), timestamp=timestamp[.N], elapsed=sum(elapsed), in_rows=in_rows[1L], out_rows=out_rows[.N]), .(dtq_id, env, src)]
   }
   return(dt)
 }